+++ library/src/main/java/com/getstream/sdk/chat/view/messageinput/AttachmentsController.kt
@@ -65,18 +65,17 @@ internal class AttachmentsController(
                     isMedia
                 )
             }
-            selectedAttachments = filterMediaFiles(isMedia, selectedAttachments)
             setSelectedAttachmentAdapter(messageInputType, isMedia)
             view.showLoadingTotalAttachments(false)
         }
     }
 
-    private fun filterMediaFiles(isMedia: Boolean, filesToFilter: Set<AttachmentMetaData>): Set<AttachmentMetaData> {
-        return if (!isMedia) {
-            filesToFilter
-        } else {
-            filesToFilter.filter(mediaAttachmentsPredicate).toSet()
-        }
+    private fun filterAttachments(
+        isMedia: Boolean,
+        filesToFilter: Set<AttachmentMetaData>
+    ): Set<AttachmentMetaData> {
+        return filesToFilter.filter(if (isMedia) mediaAttachmentsPredicate else fileAttachmentsPredicate)
+            .toSet()
     }
 
     private suspend fun getAttachmentsFromLocal(
@@ -201,10 +200,11 @@ internal class AttachmentsController(
         }
     }
 
-    private fun setSelectedAttachmentAdapter(
+    internal fun setSelectedAttachmentAdapter(
         messageInputType: MessageInputType?,
         isMedia: Boolean
     ) {
+        selectedAttachments = filterAttachments(isMedia, selectedAttachments)
         if (isMedia) {
             selectedMediaAttachmentAdapter.setAttachments(selectedAttachments.toList())
             selectedMediaAttachmentAdapter.cancelListener =
@@ -243,6 +243,16 @@ internal class AttachmentsController(
         }
     }
 
+    internal fun selectAttachmentsFromUriList(uriList: List<Uri>) {
+        GlobalScope.launch(dispatchersProvider.mainDispatcher) {
+            setSelectedAttachmentAdapter(null, false)
+            val attachments = withContext(dispatchersProvider.ioDispatcher) {
+                storageHelper.getAttachmentsFromUriList(view.context, uriList)
+            }
+            attachments.forEach { selectAttachment(it, false) }
+        }
+    }
+
     internal fun selectAttachment(
         attachment: AttachmentMetaData,
         isMedia: Boolean
@@ -265,17 +275,14 @@ internal class AttachmentsController(
             permissionChecker.isGrantedCameraPermissions(view.context) -> {
                 view.showMediaPermissions(false)
                 view.showCameraPermissions(false)
-                view.showFilePermissions(true)
             }
             permissionChecker.isGrantedStoragePermissions(view.context) -> {
                 view.showMediaPermissions(false)
                 view.showCameraPermissions(true)
-                view.showFilePermissions(true)
             }
             else -> {
                 view.showMediaPermissions(true)
                 view.showCameraPermissions(true)
-                view.showFilePermissions(true)
             }
         }
     }
@@ -296,6 +303,9 @@ internal class AttachmentsController(
 
     companion object {
         private val listOfMediaTypes = listOf(ModelType.attach_image, ModelType.attach_video)
-        private val mediaAttachmentsPredicate: (AttachmentMetaData) -> Boolean = { it.type in listOfMediaTypes }
+        private val mediaAttachmentsPredicate: (AttachmentMetaData) -> Boolean =
+            { it.type in listOfMediaTypes }
+        private val fileAttachmentsPredicate: (AttachmentMetaData) -> Boolean =
+            { it.type == ModelType.attach_file }
     }
 }