+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # To be released:
 
+- Improve selecting non-media attachments
+- Fix showing attachments captured with camera
+
 # Oct 7th, 2020 - 4.3.0-beta-4
 - For Java clients, the `bindView` methods used to bind a ViewModel and its UI component together are now available with friendlier syntax.
 - Calls such as `MessageListViewModelBindingKt.bindView(...);` should be replaced with calls like `MessageListViewModelBinding.bind(...);`